git-svn-id: svn://123.57.173.36/meijia-parent@1309 52ac4245-7803-4e87-b018-04b7c5ce731e

+++ simi-dao/src/main/java/com/simi/po/model/order/OrderExtGreen.java
@@ -1,5 +1,7 @@
 package com.simi.po.model.order;
 
+import java.math.BigDecimal;
+
 
 public class OrderExtGreen {
 	
@@ -15,7 +17,7 @@ public class OrderExtGreen {
     
     private Long totalNum;
     
-    private Long totalBudget;
+    private BigDecimal totalBudget;
     
     private Long addTime;
 
@@ -67,13 +69,11 @@ public class OrderExtGreen {
 		this.totalNum = totalNum;
 	}
 
-
-
-	public Long getTotalBudget() {
+	public BigDecimal getTotalBudget() {
 		return totalBudget;
 	}
 
-	public void setTotalBudget(Long totalBudget) {
+	public void setTotalBudget(BigDecimal totalBudget) {
 		this.totalBudget = totalBudget;
 	}
